requirements.txt
streamlit
openai
pillow 
python-dotenv
PyYAML
opencv-python